Puesto cosas del game en otro archivo
